2022 May 11 - Added broad phases selector, and a body selector.
Collision output is currently not working.

+++ Assets/Scripts/Common/UI/UIIntSlider.cs
@@ -25,7 +25,7 @@ public class UIIntSlider : MonoBehaviour
 		slider.minValue = data.min;
 		slider.maxValue = data.max;
 
-		data.value = (int)Mathf.Clamp(data.value, data.min, data.max);
+		data.value = Mathf.Clamp(data.value, data.min, data.max);
 
 		slider.onValueChanged.AddListener(UpdateValue);
 		slider.SetValueWithoutNotify(data.value);
